+++ common/common.go
@@ -20,7 +20,7 @@ const (
 )
 
 // GetSafeShare does the same as GetUncappedShare , but GetSafeShare will guarantee the result will not more than total
-func GetSafeShare(part, total, allocation cosmos.Int) cosmos.Int {
+func GetSafeShare(part, total, allocation cosmos.Dec) cosmos.Dec {
 	if part.GTE(total) {
 		part = total
 	}
@@ -29,33 +29,17 @@ func GetSafeShare(part, total, allocation cosmos.Int) cosmos.Int {
 
 // GetUncappedShare this method will panic if any of the input parameter can't be convert to cosmos.Dec
 // which shouldn't happen
-func GetUncappedShare(part, total, allocation cosmos.Int) (share cosmos.Int) {
+func GetUncappedShare(part, total, allocation cosmos.Dec) (share cosmos.Dec) {
 	if part.IsZero() || total.IsZero() {
-		return cosmos.ZeroInt()
+		return cosmos.ZeroDec()
 	}
 	defer func() {
 		if err := recover(); err != nil {
-			share = cosmos.ZeroInt()
+			share = cosmos.ZeroDec()
 		}
 	}()
-	// use string to convert cosmos.Int to cosmos.Dec is the only way I can find out without being constrain to uint64
-	// cosmos.Int can hold values way larger than uint64 , because it is using big.Int internally
-	aD, err := cosmos.NewDecFromStr(allocation.String())
-	if err != nil {
-		panic(fmt.Errorf("fail to convert %s to cosmos.Dec: %w", allocation.String(), err))
-	}
-
-	pD, err := cosmos.NewDecFromStr(part.String())
-	if err != nil {
-		panic(fmt.Errorf("fail to convert %s to cosmos.Dec: %w", part.String(), err))
-	}
-	tD, err := cosmos.NewDecFromStr(total.String())
-	if err != nil {
-		panic(fmt.Errorf("fail to convert%s to cosmos.Dec: %w", total.String(), err))
-	}
 	// A / (Total / part) == A * (part/Total) but safer when part < Totals
-	result := aD.Quo(tD.Quo(pD))
-	share = cosmos.NewIntFromBigInt(result.RoundInt().BigInt())
+	share = allocation.Quo(total.Quo(part))
 	return
 }