+++ scripts/genesis.sh
@@ -64,6 +64,27 @@ add_claim_records() {
 	mv /tmp/genesis.json ~/.arkeo/config/genesis.json
 }
 
+set_fee_pool() {
+    local denom="$1"
+    local amount="$2"
+
+    jq --arg DENOM "$denom" --arg AMOUNT "$amount" '.app_state.distribution.fee_pool.community_pool = [{
+        "denom": $DENOM,
+        "amount": $AMOUNT
+    }]' <~/.arkeo/config/genesis.json >/tmp/genesis.json
+    mv /tmp/genesis.json ~/.arkeo/config/genesis.json
+}
+
+disable_mint_params() {
+    jq '.app_state.mint.minter.inflation = "0.000000000000000000" |
+        .app_state.mint.minter.annual_provisions = "0.000000000000000000" |
+        .app_state.mint.params.inflation_rate_change = "0.000000000000000000" |
+        .app_state.mint.params.inflation_max = "0.000000000000000000" |
+        .app_state.mint.params.inflation_min = "0.000000000000000000"' \
+    <~/.arkeo/config/genesis.json >/tmp/genesis.json
+    mv /tmp/genesis.json ~/.arkeo/config/genesis.json
+}
+
 if [ ! -f ~/.arkeo/config/priv_validator_key.json ]; then
 	# remove the original generate genesis file, as below will init chain again
 	rm -rf ~/.arkeo/config/genesis.json
@@ -79,28 +100,24 @@ if [ ! -f ~/.arkeo/config/genesis.json ]; then
 
 	arkeod keys add faucet --keyring-backend test
 	FAUCET=$(arkeod keys show faucet -a --keyring-backend test)
-	add_account "$FAUCET" $TOKEN 10000000000000000 # faucet, 10m
+	add_account "$FAUCET" $TOKEN 2900000000000000 # faucet, 29m
+	disable_mint_params
 
 	if [ "$NET" = "mocknet" ] || [ "$NET" = "testnet" ]; then
-		add_module tarkeo14tmx70mvve3u7hfmd45vle49kvylk6s2wllxny $TOKEN 30250000000000 'claimarkeo'
+		add_module tarkeo1d0m97ywk2y4vq58ud6q5e0r3q9khj9e3unfe4t $TOKEN 2420000000000000 'arkeo-reserve' 
+		add_module tarkeo14tmx70mvve3u7hfmd45vle49kvylk6s2wllxny $TOKEN 3025000000000000 'claimarkeo'
+		add_module tarkeo1jv65s3grqf6v6jl3dp4t6c9t9rk99cd8t6gr9e $TOKEN 4840000000000000 'distribution'
+		# this is to handle the balance set to distribution as a community pool 
+		set_fee_pool $TOKEN 4840000000000000
 
-		echo "shoulder heavy loyal save patient deposit crew bag pull club escape eyebrow hip verify border into wire start pact faint fame festival solve shop" | arkeod keys add alice --keyring-backend test --recover
-		ALICE=$(arkeod keys show alice -a --keyring-backend test)
-		add_account "$ALICE" $TOKEN 1100000000000000 # alice, 1.1m
+		
 
-		echo "clog swear steak glide artwork glory solution short company borrow aerobic idle corn climb believe wink forum destroy miracle oak cover solid valve make" | arkeod keys add bob --keyring-backend test --recover
-		BOB=$(arkeod keys show bob -a --keyring-backend test)
-		add_account "$BOB" $TOKEN 1000000000000000 # bob, 1m
-		add_claim_records "ARKEO" "$BOB" 1000 1000 1000 true
-
-
-		# Add Foundational Accounts
-		#  FoundationCommunityAccount = "tarkeo124qmjmg55v6q5c5vy0vcpefrywxnxhkm7426pc"
-		add_account "tarkeo124qmjmg55v6q5c5vy0vcpefrywxnxhkm7426pc" $TOKEN 1048400000000000000
+		# Add Foundational Accounts            
 		# 	FoundationDevAccount       = "tarkeo1x978nttd8vgcgnv9wxut4dh7809lr0n2fhuh0q"
-		add_account "tarkeo1x978nttd8vgcgnv9wxut4dh7809lr0n2fhuh0q" $TOKEN 12100000000000
+		add_account "tarkeo1x978nttd8vgcgnv9wxut4dh7809lr0n2fhuh0q" $TOKEN 1210000000000000
 		#   FoundationGrantsAccount    = "tarkeo1a307z4a82mcyv9njdj9ajnd9xpp90kmeqwntxj"
-		add_account "tarkeo1a307z4a82mcyv9njdj9ajnd9xpp90kmeqwntxj" $TOKEN 6050000000000
+		add_account "tarkeo1a307z4a82mcyv9njdj9ajnd9xpp90kmeqwntxj" $TOKEN 605000000000000
+		                                                                   
 	
 	
 		# Thorchain derived test addresses